pyproject.toml
[project]
name = "gex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "faiss-cpu>=1.7.4",
    "fastapi>=0.109.0",
    "httpx>=0.28.1",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-ollama>=0.0.1",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.0.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "streamlit>=1.32.0",
    "uvicorn>=0.27.0",
    "dotenv>=0.9.9",
    "duckduckgo-search==8.0.1",
    "docx2txt",
    "pypdf",
    "duckduckgo-search",
]
